# Conflicts:
#	core/src/main/java/se2/ticktackbumm/core/TickTackBummGame.java
#	core/src/main/java/se2/ticktackbumm/core/client/ClientMessageHandler.java
#	core/src/main/java/se2/ticktackbumm/core/models/BombImpl/Bomb.java
#	core/src/main/java/se2/ticktackbumm/core/models/BombImpl/BombExplosion.java
#	core/src/main/java/se2/ticktackbumm/core/network/messages/client/StartBomb.java
#	core/src/main/java/se2/ticktackbumm/core/screens/MainGameScreen.java
#	core/src/main/java/se2/ticktackbumm/core/screens/SpinWheelScreen.java
#	server/src/main/java/se2/ticktackbumm/server/network/ServerMessageHandler.java
#	server/src/main/java/se2/ticktackbumm/server/network/ServerMessageSender.java

+++ core/src/main/java/se2/ticktackbumm/core/TickTackBummGame.java
@@ -144,7 +144,7 @@ public class TickTackBummGame extends Game {
         // TODO: show game finished message, scoreboard screen, ...
     }
 
-    public void setBombToTick(float timer){
+    public void setBombToTick(int timer){
         MainGameScreen gameScreen = (MainGameScreen) this.getScreen();
 
         gameScreen.getBomb().setExplodeTime(timer);
